+++ application/Sourcecode/Enemy/IEnemy.cpp
@@ -69,7 +69,7 @@ void IEnemy::ColPosUpdate()
 	col_.radius = obj_->GetTransform()->scale_.x;
 }
 
-void IEnemy::Damage(const Vector3& knockVec, int32_t damageValue, int32_t cooltime)
+void IEnemy::Damage(const Vector3& knockVec, int32_t damageValue, float cooltime)
 {
 	if (damageCoolTime_.GetIsEnd())
 	{
